PROJFBPRO-69: Refactor functions into mdr_event_information_extraction_functions.py
