--- src/get_mdr_name_from_event.py
@@ -1,13 +1,0 @@
-import logging
-
-
-# Returns the name of the attacked system
-def get_mdr_sys_name(mdr_event) -> str:
-    sys_name = None
-    try:
-        sys_name = mdr_event['event']['taxonomy']['External']['subject']['machine']['Local']['host']['name']
-    except KeyError:
-        logging.error("Error in key path. Compare path in json file and in variable declaration!")
-        raise
-
-    return sys_name
