+++ test/TestExtractMDRName.py
@@ -1,7 +1,7 @@
 import unittest
 import json
 
-from src.get_mdr_name_from_event import get_mdr_sys_name
+from src.mdr_event_information_extraction_functions import get_mdr_sys_name
 
 
 class TestExtractMDRName(unittest.TestCase):
